+++ packages/captcha-nodejs/src/errors.ts
@@ -1,9 +1,7 @@
-import SuperTokensError from "supertokens-node/lib/build/error";
-
 type CaptchaPluginErrorType = "CAPTCHA_VERIFICATION_ERROR" | "PLUGIN_CONFIG_ERROR";
 
-export class CaptchaPluginError extends SuperTokensError {
-  constructor(type: CaptchaPluginErrorType, message: string) {
-    super({ type, message });
+export class CaptchaPluginError extends Error {
+  constructor(public type: CaptchaPluginErrorType, message: string) {
+    super(message)
   }
 }