+++ packages/profile-base-react/src/plugin.ts
@@ -19,6 +19,7 @@ const { usePluginContext, setContext } = buildContext<{
   pluginConfig: SuperTokensPluginProfileConfig;
   getSections: () => SuperTokensPluginProfileSection[];
   registerSection: RegisterSection;
+  getOnLoadHandlers: () => (() => Promise<void>)[];
   t: (key: TranslationKeys, replacements?: Record<string, string>) => string;
 }>();
 export { usePluginContext };
@@ -34,6 +35,14 @@ export const init = createPluginInitFunction<
 
     const getSections = () => {
       return sections.sort((a, b) => a.order - b.order);
+    };
+
+    const onLoadHandlers: (() => Promise<void>)[] = [];
+    const registerOnLoadHandler = (handler: () => Promise<void>) => {
+      onLoadHandlers.push(handler);
+    };
+    const getOnLoadHandlers = () => {
+      return onLoadHandlers;
     };
 
     let sectionOrder = 0;
@@ -84,6 +93,7 @@ export const init = createPluginInitFunction<
           pluginConfig,
           getSections,
           registerSection,
+          getOnLoadHandlers,
           t,
         });
       },
@@ -101,6 +111,7 @@ export const init = createPluginInitFunction<
       exports: {
         getSections,
         registerSection,
+        registerOnLoadHandler,
       },
     };
   },