+++ packages/progressive-profiling-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @supertokens-plugins/progressive-profiling-react
 
+## 0.0.5
+
+### Patch Changes
+
+- 3adf48f: Added documenation
+
 ## 0.0.4
 
 ### Patch Changes